+++ app/Http/Controllers/Struktur/PositionController.php
@@ -21,6 +21,9 @@ class PositionController extends Controller
     }
     public function index()
     {
+        if(!auth()->user()->hasRole('Admin')){
+            return back()->with('error','Anda Tidak Mempunyai Akses Halaman Ini!');
+        }
         return view('struktur.position.index');
     }
 
@@ -31,6 +34,9 @@ class PositionController extends Controller
      */
     public function create()
     {
+        if(!auth()->user()->hasRole('Admin')){
+            return back()->with('error','Anda Tidak Mempunyai Akses Halaman Ini!');
+        }
         return view('struktur.position.create');
 
     }
@@ -73,6 +79,9 @@ class PositionController extends Controller
      */
     public function edit($id)
     {
+        if(!auth()->user()->hasRole('Admin')){
+            return back()->with('error','Anda Tidak Mempunyai Akses Halaman Ini!');
+        }
         $oldData = Position::find($id);
         return view('struktur.position.edit',compact('oldData'));
 
@@ -106,6 +115,9 @@ class PositionController extends Controller
      */
     public function destroy($id)
     {
+        if(!auth()->user()->hasRole('Admin')){
+            return back()->with('error','Anda Tidak Mempunyai Akses Halaman Ini!');
+        }
         $deleteData = Position::find($id);
         $deleteData->delete();
         $this->sinkron_status();
